# api/openapi.yaml
openapi: 3.1.0
info:
  title: NXT TIME PULSE Admin API
  version: 1.0.0
  description: Admin-gated API endpoints for GEO TIMESHEET - TIMEPULSE
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
paths:
...
  /users/add:
    post:
      summary: Add existing Clerk user by email (admin/manager only)
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [admin, manager, worker] }
      responses:
        "200":
          description: User added
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found in Clerk
        "409":
          description: User already member
...
  /users/{id}/timeline:
    get:
      summary: User lifecycle timeline, assignment history, and pending requests
      tags: [Users]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User timeline payload
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Unauthorized" }
    Forbidden:
      description: Insufficient permissions (requires org:admin or org:manager)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Forbidden" }